+++ backend/src/test/kotlin/com/example/backend2/unitTest/domain/winner/entity/WinnerTest.kt
@@ -9,8 +9,10 @@ import com.example.backend2.domain.winner.entity.Winner
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.springframework.test.context.ActiveProfiles
 import java.time.LocalDateTime
 
+@ActiveProfiles("test")
 @DisplayName("Winner 엔티티 단위 테스트")
 class WinnerTest {
     @Test
